requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
openai
faiss-cpu
gtts
speechrecognition
pydub 